LayoutTests/platform/chromium/http/tests/security/xss-DENIED-frame-name-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://localhost:8000" from accessing a frame with origin "http://127.0.0.1:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 2: Uncaught TypeError: Cannot read property 'prop' of undefined
   This test passes if it doesn't alert fail.
